Reduce coupling of parser
By splitting up the parser into traits, the coupling can be reduced to
below 13.
